refactor(database): split storage responsibilities and repositories
