+++ src-tauri/src/infrastructure/db/repositories/mod.rs
@@ -1,0 +1,14 @@
+//! 业务仓储：按业务域组织 user_sessions / auth_cookie_v2 /
+//! custom_schedule_courses / online_learning / chaoxing_checkin_log 的读写。
+
+pub mod auth_cookie;
+pub mod chaoxing;
+pub mod custom_schedule;
+pub mod online_learning;
+pub mod session;
+
+pub use auth_cookie::*;
+pub use chaoxing::*;
+pub use custom_schedule::*;
+pub use online_learning::*;
+pub use session::*;
